requirements.txt
celery>=4.2.1
connexion>=2.2.0
connexion-buzz>=0.2.1
dynaconf>=2.0.3
flask-marshmallow==0.10.1
infi.clickhouse_orm==1.2.0
marshmallow==3.3.0
redis>=3.2.0
requests>=2.21.0
sentry-sdk[flask]>=0.7.10
uWSGI>=2.0.18